Vi/xdg conformant config (#96)
* Conform to the XDG specification for configuration file storage.

This commit modifies the location of the `.entropy-cli.config` file on
systems that conform to the [XDG
specification](https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html).
Specifically, it checks for the existence of the `XDG_CONFIG_HOME`
environment variable and, if present, appends an Entropy-specific
directory name to that path to use as the base filesystem path for where
to find configuration information. If that environment variable is not
set, this patch falls back to the original location of the user's
`$HOME` directory.

THe XDG specification expects `XDG_CONFIG_HOME`, if present, to be set
to `$HOME/.config`, a directory that should already exist on conformant
systems. Inside that directory, packages and vendors are expected to
place user-specific configuration data. In our case, the directory
`entropy-cryptography` is created inside `$XDG_CONFIG_HOME` if it is not
already present. If it is present, we notice its existence and simply
proceed as normal.

The rationale for this change is straightforward and is as follows:

1. Some other internal Entropy systems are already XDG-conformant. This
   would help ensure all Entropy-related tooling is stored in
   predictable locations that are all maintained in a single directory.
1. Modern GNU/Linux desktop environments are already set up to expect
   this, making it easier to integrate with, e.g., backup and system
   tooling on end-user devices.
1. Shows that we are a good desktop citizen, avoiding clutter in the
   user's `$HOME` directory.

* add mkdirp, env-paths

* move init

* fix config path

---------

Co-authored-by: Violet Parr <vi@entropy.xyz>